Move provenance adding code to utility package
